ml_backend/requirements.txt
fastapi
uvicorn
ultralytics
torch
torchvision
numpy
pillow
python-multipart
cors 